Move IFuture definition to i_task_specification module
