+++ traits_futures/api.py
@@ -108,9 +108,8 @@ from traits_futures.future_states import (
     WAITING,
 )
 from traits_futures.i_event_loop import IEventLoop
-from traits_futures.i_future import IFuture
 from traits_futures.i_parallel_context import IParallelContext
-from traits_futures.i_task_specification import ITaskSpecification
+from traits_futures.i_task_specification import IFuture, ITaskSpecification
 from traits_futures.multiprocessing_context import MultiprocessingContext
 from traits_futures.multithreading_context import MultithreadingContext
 from traits_futures.traits_executor import TraitsExecutor
